Made the Movement class independent from Cursor class
All methods in Movement now take a Cursor object from which they extract
the necessary info to function and gives back a positive or negative
movement integer that is translated by Cursor to a Vector position in
the matrix. I am unsure if this is the best thing for the future, as I
want the Movement-logic to be available for more objects inside the game.