Added graphical reaction between cursor and letters
Cell now has a default and a temporary texture with methods for manipulating the two.
Now when a cursor is over a white_green letter it will change
the font to white_purple. This is right now just hardcoded into
Cursor, but will be turned into a general method with a default color
scheme